+++ src/babel/root-visitor.ts
@@ -70,7 +70,6 @@ export default function rootVisitor(
             mode: "compileAndTransform",
             blockModuleTransform: [],
             hasStyledComponentImport: false,
-            hasClassNames: false,
             canCompile,
             canTransform,
             ...state,
@@ -92,22 +91,19 @@ export default function rootVisitor(
             hasStyleSheetImport,
             hasProvider,
             hasStyledComponentImport,
-            hasClassNames,
             hmr,
           } = visitorState;
 
           if (hmr) {
-            // There are no classNames so skip this file
-            if (!hasClassNames) {
-              return;
-            }
-
             /**
              * Override tailwind to only process the classnames in this file
              */
             const { styles } = extractStyles({
               ...tailwindConfig,
               content: [filename],
+              // If the file doesn't have any Tailwind styles, it will print a warning
+              // We force an empty style to prevent this
+              safelist: [".native-hmr-empty"],
             });
 
             const bodyNode = path.node.body;