+++ src/with-class-names.ts
@@ -1,5 +1,7 @@
 export interface WithClassNames {
+  baseClassName?: string;
   className?: string;
+  baseTw?: string;
   twClassName?: string;
   propsToTransform?: string[];
   componentProps: Record<string, unknown>;
@@ -11,14 +13,20 @@ const isComponentRegex = /(?:^|\s)(component)(?:$|\s)/gi;
 const isParentRegex = /(?:^|\s)(parent)(?:$|\s)/gi;
 
 export function withClassNames({
+  baseClassName,
   className,
+  baseTw,
   twClassName,
   componentProps,
   propsToTransform = [],
   spreadProps = [],
   classProps = [],
 }: WithClassNames) {
-  const classes = twClassName ?? className ?? "";
+  const classes = [
+    baseTw ?? baseClassName ?? "",
+    twClassName ?? className ?? "",
+  ].join(" ");
+
   const isComponent = isComponentRegex.test(classes);
   const isParent = isParentRegex.test(classes);